Bugfix/manual update cypress 10 (#665)
* Update package.json

* Create cypress.conf.ts

* Rename cypress.conf.ts to cypress.config.ts

* [AUTO] prettier format write

* Rename cypress.json to old-cypress-json

* Update package.json

* Update package.json

* [AUTO] prettier format write

* Update package.json

* Update tsconfig.json

* [AUTO] prettier format write

* Update tsconfig.json

* Rename client/cypress.config.ts to client/src/ cypress.config.ts

* Update cypress.config.ts

* Update package.json

* change to js config

* Update package.json
