+++ client/package.json
@@ -7,8 +7,8 @@
     "build": "webpack --node-env production --progress --color --env PUBLIC_PATH=\"'.'\"",
     "build:doc": "styleguidist build",
     "start": "webpack serve --open --progress --color --node-env development",
-    "cy:open": "cypress open",
-    "cy:run": "wait-on http://localhost:3000/ && cypress run --config-file cypress.json",
+    "cy:open": "cypress open --config-file=cypress.config.js",
+    "cy:run": "wait-on http://localhost:3000/ && cypress run --config-file=cypress.config.js",
     "prettier": "prettier --write . --ignore-path ./.gitignore",
     "lint:eslint": "eslint . --fix --format html --output-file ./dist/reports/lint/eslint-report.html --ignore-path ./.gitignore",
     "lint:eslint:json": "eslint . --fix --format json --output-file ./dist/reports/lint/eslint-report.json --ignore-path ./.gitignore",
@@ -35,6 +35,7 @@
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
     "@cypress/code-coverage": "^3.10.0",
+    "@cypress/webpack-dev-server": "^2.0.0",
     "@gfx/zopfli": "^1.0.15",
     "@mguillaumef/audiso": "^0.0.4",
     "@testing-library/jest-dom": "^5.16.4",
@@ -59,7 +60,7 @@
     "compression-webpack-plugin": "^10.0.0",
     "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^4.0.0",
-    "cypress": "^9.7.0",
+    "cypress": "^10.3.0",
     "cypress-movie": "^3.0.20",
     "cypress-multi-reporters": "^1.6.1",
     "cypress-sonarqube-reporter": "^1.10.0",
